add methods to support adding a new member to a Cloudflare account; moved DTOs to separate artifact
