+++ build.sbt
@@ -1,6 +1,7 @@
+import Dependencies._
+
 lazy val buildSettings = Seq(
   organization := "com.dwolla",
-  name := "cloudflare-api-client",
   homepage := Some(url("https://github.com/Dwolla/scala-cloudflare")),
   description := "Scala library for the Cloudflare v4 API",
   licenses += ("MIT", url("http://opensource.org/licenses/MIT")),
@@ -14,21 +15,28 @@ lazy val buildSettings = Seq(
   resolvers ++= Seq(
     Resolver.bintrayIvyRepo("dwolla", "maven")
   ),
-  libraryDependencies ++= {
-    val specs2Version = "3.8.6"
-    val json4sVersion = "3.5.1"
-    Seq(
-      "com.jsuereth" %% "scala-arm" % "2.0",
-      "org.json4s" %% "json4s-native" % json4sVersion,
-      "org.apache.httpcomponents" % "httpclient" % "4.5.2",
-      "org.specs2" %% "specs2-core" % specs2Version % Test,
-      "org.specs2" %% "specs2-mock" % specs2Version % Test,
-      "org.specs2" %% "specs2-matcher-extra" % specs2Version % Test,
-      "com.dwolla" %% "testutils" % "1.4.0" % Test
-    )
-  },
   scalacOptions += "-deprecation"
 )
+
+lazy val dto = (project in file("dto"))
+  .settings(buildSettings: _*)
+  .settings(name := "cloudflare-api-dto")
+
+lazy val client = (project in file("client"))
+  .settings(buildSettings: _*)
+  .settings(name := "cloudflare-api-client")
+  .settings(libraryDependencies ++= Seq(
+      scalaArm,
+      json4s,
+      httpComponents,
+      specs2Core % Test,
+      specs2Mock % Test,
+      specs2Matchers % Test,
+      dwollaTestUtils % Test
+    )
+  )
+  .dependsOn(dto)
+
 
 lazy val bintraySettings = Seq(
   bintrayVcsUrl := homepage.value.map(_.toString),
@@ -38,5 +46,6 @@ lazy val bintraySettings = Seq(
   pomIncludeRepository := { _ ⇒ false }
 )
 
-lazy val root = (project in file("."))
+lazy val scalaCloudflare = (project in file("."))
   .settings(buildSettings ++ bintraySettings: _*)
+  .aggregate(dto, client)
