+++ client/src/main/scala/com/dwolla/cloudflare/domain/model/Error.scala
