+++ client/src/main/scala/com/dwolla/cloudflare/domain/model/accounts/Account.scala
@@ -1,0 +1,7 @@
+package com.dwolla.cloudflare.domain.model.accounts
+
+case class Account (
+  id: String,
+  name: String,
+  settings: AccountSettings
+)
