+++ client/src/main/scala/com/dwolla/cloudflare/domain/model/accounts/AccountMember.scala
@@ -1,0 +1,8 @@
+package com.dwolla.cloudflare.domain.model.accounts
+
+case class AccountMember (
+  id: String,
+  user: User,
+  status: String,
+  roles: List[AccountRole]
+)
