+++ client/src/main/scala/com/dwolla/cloudflare/domain/model/accounts/AccountSettings.scala
@@ -1,0 +1,5 @@
+package com.dwolla.cloudflare.domain.model.accounts
+
+case class AccountSettings (
+  enforceTwoFactor: Boolean
+)
