+++ client/src/main/scala/com/dwolla/cloudflare/domain/model/accounts/User.scala
@@ -1,0 +1,9 @@
+package com.dwolla.cloudflare.domain.model.accounts
+
+case class User (
+  id: String,
+  firstName: String,
+  lastName: String,
+  emailAddress: String,
+  twoFactorEnabled: Boolean
+)
