+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/JsonWritable.scala
@@ -1,0 +1,3 @@
+package com.dwolla.cloudflare.domain.dto
+
+trait JsonWritable extends AnyRef
