+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/accounts/AccountDTO.scala
@@ -1,0 +1,7 @@
+package com.dwolla.cloudflare.domain.dto.accounts
+
+case class AccountDTO (
+  id: String,
+  name: String,
+  settings: AccountSettingsDTO
+)
