+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/accounts/AccountMemberDTO.scala
@@ -1,0 +1,8 @@
+package com.dwolla.cloudflare.domain.dto.accounts
+
+case class AccountMemberDTO (
+  id: String,
+  user: UserDTO,
+  status: String,
+  roles: List[AccountRoleDTO]
+)
