+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/accounts/AccountRoleDTO.scala
@@ -1,0 +1,13 @@
+package com.dwolla.cloudflare.domain.dto.accounts
+
+case class AccountRoleDTO (
+  id: String,
+  name: String,
+  description: String,
+  permissions: Map[String, AccountRolePermissionsDTO]
+)
+
+case class AccountRolePermissionsDTO (
+  read: Boolean,
+  edit: Boolean
+)
