+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/accounts/AccountSettingsDTO.scala
@@ -1,0 +1,5 @@
+package com.dwolla.cloudflare.domain.dto.accounts
+
+case class AccountSettingsDTO (
+  enforce_twofactor: Boolean
+)
