+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/accounts/NewAccountMemberDTO.scala
@@ -1,0 +1,9 @@
+package com.dwolla.cloudflare.domain.dto.accounts
+
+import com.dwolla.cloudflare.domain.dto.JsonWritable
+
+case class NewAccountMemberDTO (
+  email: String,
+  roles: List[String],
+  status: Option[String] = None
+) extends JsonWritable
