+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/accounts/UserDTO.scala
@@ -1,0 +1,9 @@
+package com.dwolla.cloudflare.domain.dto.accounts
+
+case class UserDTO (
+  id: String,
+  first_name: String,
+  last_name: String,
+  email: String,
+  two_factor_authentication_enabled: Boolean
+)
