+++ dto/src/main/scala/com/dwolla/cloudflare/domain/dto/dns/DnsRecordDTO.scala
@@ -1,0 +1,10 @@
+package com.dwolla.cloudflare.domain.dto.dns
+
+import com.dwolla.cloudflare.domain.dto.JsonWritable
+
+case class DnsRecordDTO(id: Option[String] = None,
+                        name: String,
+                        content: String,
+                        `type`: String,
+                        ttl: Option[Int] = None,
+                        proxied: Option[Boolean] = None) extends JsonWritable
